+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Added `@futagoza/create-stream`, a non-gulp package
 * Added `@futagoza/pump`, an optional gulp package
 * Added `@futagoza/child-process`, a non-gulp package
+* Set `@futagoza/gulp` to use `@futagoza/child-process`
+* Set `@futagoza/node-run` to use `@futagoza/child-process`
+* Set `@futagoza/publish-package` to use `@futagoza/child-process`
 
 ## [v0.5.1](https://github.com/futagoza/gulp/compare/v0.5.0...v0.5.1) _(2018-09-26)_